// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const PartnerType = {
  RESELLER: 'RESELLER',
  END_USER: 'END_USER'
} as const

export type PartnerType = (typeof PartnerType)[keyof typeof PartnerType]


export const PaymentModel = {
  FREE: 'FREE',
  ONETIME: 'ONETIME',
  RECURRING: 'RECURRING'
} as const

export type PaymentModel = (typeof PaymentModel)[keyof typeof PaymentModel]


export const FieldType = {
  TEXT: 'TEXT',
  TEXTAREA: 'TEXTAREA',
  NUMBER: 'NUMBER',
  PRICE: 'PRICE',
  DROPDOWN: 'DROPDOWN',
  CHECKBOX: 'CHECKBOX',
  DATE: 'DATE',
  DATETIME: 'DATETIME',
  LINK: 'LINK',
  TABLE: 'TABLE',
  ATTACHMENT: 'ATTACHMENT'
} as const

export type FieldType = (typeof FieldType)[keyof typeof FieldType]


export const PricingModel = {
  FIXED: 'FIXED',
  DISCOUNT: 'DISCOUNT',
  TIERED: 'TIERED'
} as const

export type PricingModel = (typeof PricingModel)[keyof typeof PricingModel]


export const OrderMode = {
  DIRECT: 'DIRECT',
  REQUEST: 'REQUEST'
} as const

export type OrderMode = (typeof OrderMode)[keyof typeof OrderMode]